Moved metric calculation function to metrics.calculator but will nedd further work
