+++ solarforecastarbiter/metrics/tests/test_calculator.py
@@ -2,7 +2,7 @@ import pandas as pd
 import pytest
 
 
-from solarforecastarbiter.reports import metrics
+from solarforecastarbiter.metrics import calculator
 
 
 @pytest.mark.parametrize('fx,obs', [
@@ -21,4 +21,4 @@ from solarforecastarbiter.reports import metrics
 ])
 def test_calculate_metrics_runs(report_objects, fx, obs):
     fxobs = report_objects[0].forecast_observations[0]
-    metrics.calculate_metrics(fxobs, fx, obs)
+    calculator.calculate_metrics(fxobs, fx, obs)
